# openapi.yaml
openapi: 3.1.0
info:
  description: LensBridge backend — media upload and MusallahBoard management
  title: LensBridge API
  version: 0.0.1-SNAPSHOT
servers:
- url: /
security:
- bearerAuth: []
paths:
  /api/admin/audit:
    get:
      operationId: getAuditEvents
      parameters:
      - description: Zero-based page index (0..N)
        in: query
        name: page
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      - description: The size of the page to be returned
        in: query
        name: size
        required: false
        schema:
          type: integer
          default: 20
          minimum: 1
      - description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        in: query
        name: sort
        required: false
        schema:
          type: array
          items:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageAuditEventDto"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: Request failed; body carries a human-readable message
      tags:
      - admin-controller
  /api/admin/audit/action/{action}:
    get:
      operationId: getAuditEventsByAction
      parameters:
      - in: path
        name: action
        required: true
        schema:
          type: string
          enum:
          - APPROVE_UPLOAD
          - UNAPPROVE_UPLOAD
          - DELETE_UPLOAD
          - FEATURE_UPLOAD
          - UNFEATURE_UPLOAD
          - CREATE_EVENT
          - UPDATE_EVENT
          - DELETE_EVENT
          - CREATE_CALENDAR_EVENT
          - UPDATE_CALENDAR_EVENT
          - DELETE_CALENDAR_EVENT
          - CREATE_POSTER
          - UPDATE_POSTER
          - DELETE_POSTER
          - CREATE_SOCIAL
          - UPDATE_SOCIAL
          - DELETE_SOCIAL
          - SAVE_WEEKLY_CONTENT
          - DELETE_WEEKLY_CONTENT
          - UPDATE_BOARD_CONFIG
          - UPDATE_BOARD_TICKER
          - REFRESH_BOARDS
          - ISSUE_ENROLLMENT_TOKEN
          - REVOKE_DEVICE
          - ISSUE_DEVICE_COMMAND
          - PROMOTE_USER
          - DEMOTE_USER
          - DISABLE_USER
          - ENABLE_USER
          - VIEW_AUDIT_LOGS
          - EXPORT_DATA
          - SYSTEM_MAINTENANCE
          - VERIFY_USER
          - UNVERIFY_USER
          - RESET_USER_PASSWORD
          - TRIGGER_PASSWORD_RESET_EMAIL
          - ADD_USER_ROLE
          - REMOVE_USER_ROLE
          - GRANT_PERMISSION
          - REVOKE_PERMISSION
          - ADD_USER
          - REMOVE_USER
          - UPDATE_USER
      - description: Zero-based page index (0..N)
        in: query
        name: page
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      - description: The size of the page to be returned
        in: query
        name: size
        required: false
        schema:
          type: integer
          default: 20
          minimum: 1
      - description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        in: query
        name: sort
        required: false
        schema:
          type: array
          items:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageAuditEventDto"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: Request failed; body carries a human-readable message
      tags:
      - admin-controller
  /api/admin/audit/actions:
    get:
      operationId: getAvailableAuditActions
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                  enum:
                  - APPROVE_UPLOAD
                  - UNAPPROVE_UPLOAD
                  - DELETE_UPLOAD
                  - FEATURE_UPLOAD
                  - UNFEATURE_UPLOAD
                  - CREATE_EVENT
                  - UPDATE_EVENT
                  - DELETE_EVENT
                  - CREATE_CALENDAR_EVENT
                  - UPDATE_CALENDAR_EVENT
                  - DELETE_CALENDAR_EVENT
                  - CREATE_POSTER
                  - UPDATE_POSTER
                  - DELETE_POSTER
                  - CREATE_SOCIAL
                  - UPDATE_SOCIAL
                  - DELETE_SOCIAL
                  - SAVE_WEEKLY_CONTENT
                  - DELETE_WEEKLY_CONTENT
                  - UPDATE_BOARD_CONFIG
                  - UPDATE_BOARD_TICKER
                  - REFRESH_BOARDS
                  - ISSUE_ENROLLMENT_TOKEN
                  - REVOKE_DEVICE
                  - ISSUE_DEVICE_COMMAND
                  - PROMOTE_USER
                  - DEMOTE_USER
                  - DISABLE_USER
                  - ENABLE_USER
                  - VIEW_AUDIT_LOGS
                  - EXPORT_DATA
                  - SYSTEM_MAINTENANCE
                  - VERIFY_USER
                  - UNVERIFY_USER
                  - RESET_USER_PASSWORD
                  - TRIGGER_PASSWORD_RESET_EMAIL
                  - ADD_USER_ROLE
                  - REMOVE_USER_ROLE
                  - GRANT_PERMISSION
                  - REVOKE_PERMISSION
                  - ADD_USER
                  - REMOVE_USER
                  - UPDATE_USER
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: Request failed; body carries a human-readable message
      summary: List the audit action types available for filtering
      tags:
      - admin-controller
  /api/admin/audit/daterange:
    get:
      operationId: getAuditEventsByDateRange
      parameters:
      - in: query
        name: start
        required: true
        schema:
          type: string
          format: date-time
      - in: query
        name: end
        required: true
        schema:
          type: string
          format: date-time
      - description: Zero-based page index (0..N)
        in: query
        name: page
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      - description: The size of the page to be returned
        in: query
        name: size
        required: false
        schema:
          type: integer
          default: 20
          minimum: 1
      - description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        in: query
        name: sort
        required: false
        schema:
          type: array
          items:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageAuditEventDto"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: Request failed; body carries a human-readable message
      tags:
      - admin-controller
  /api/admin/audit/failed:
    get:
      operationId: getFailedOperations
      parameters:
      - description: Zero-based page index (0..N)
        in: query
        name: page
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      - description: The size of the page to be returned
        in: query
        name: size
        required: false
        schema:
          type: integer
          default: 20
          minimum: 1
      - description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        in: query
        name: sort
        required: false
        schema:
          type: array
          items:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageAuditEventDto"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: Request failed; body carries a human-readable message
      tags:
      - admin-controller
  /api/admin/audit/upload/{uploadId}:
    get:
      operationId: getUploadAuditHistory
      parameters:
      - in: path
        name: uploadId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditEventDto"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: Request failed; body carries a human-readable message
      tags:
      - admin-controller
...
  /api/admin/create-event:
    post:
      operationId: createMediaEvent
      parameters:
      - in: query
        name: eventName
        required: true
        schema:
          type: string
      - in: query
        name: eventDate
        required: true
        schema:
          type: string
          format: date-time
      - in: query
        name: status
        required: false
        schema:
          type: string
          enum:
          - UPCOMING
          - ONGOING
          - PAST
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: Request failed; body carries a human-readable message
      summary: Create a media event
      tags:
      - admin-controller
  /api/admin/events:
    get:
      operationId: getAllEventsForAdmin
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MediaEvent"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: Request failed; body carries a human-readable message
      summary: "List every event, including unpublished ones"
      tags:
      - admin-controller
  /api/admin/feature-upload/{uploadId}:
    post:
      operationId: featureUpload
      parameters:
      - in: path
        name: uploadId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: Request failed; body carries a human-readable message
      tags:
      - admin-controller
  /api/admin/permissions:
    get:
      operationId: getAvailablePermissions
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PermissionResponse"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: Request failed; body carries a human-readable message
      summary: "List every permission, for building a direct-grant picker"
      tags:
      - admin-controller
  /api/admin/roles:
    get:
      operationId: getAvailableRoles
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RoleDefinitionResponse"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: Request failed; body carries a human-readable message
      summary: List assignable roles with the permissions each one confers
      tags:
      - admin-controller
  /api/admin/upload/{uploadId}:
    delete:
      operationId: deleteUpload
      parameters:
      - in: path
        name: uploadId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: Request failed; body carries a human-readable message
      tags:
      - admin-controller
    post:
      operationId: approveUpload
      parameters:
      - in: path
        name: uploadId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: Request failed; body carries a human-readable message
      tags:
      - admin-controller
  /api/admin/upload/{uploadId}/approval:
    delete:
      operationId: unapproveUpload
      parameters:
      - in: path
        name: uploadId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: Request failed; body carries a human-readable message
      tags:
      - admin-controller
  /api/admin/upload/{uploadId}/featured:
    delete:
      operationId: unfeatureUpload
      parameters:
      - in: path
        name: uploadId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: Request failed; body carries a human-readable message
      tags:
      - admin-controller
...
  /api/events:
    get:
      operationId: getPublicEvents
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MediaEvent"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: Request failed; body carries a human-readable message
      summary: List publicly visible events
      tags:
      - events-controller
...
components:
  schemas:
...
    MediaEvent:
      type: object
      properties:
        date:
          type: string
          format: date-time
        id:
          type: string
          format: uuid
        name:
          type: string
        status:
          type: string
          enum:
          - UPCOMING
          - ONGOING
          - PAST
    MessageResponse:
      type: object
      properties:
        message:
          type: string
...
    PageAuditEventDto:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/AuditEventDto"
        empty:
          type: boolean
        first:
          type: boolean
        last:
          type: boolean
        number:
          type: integer
          format: int32
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: "#/components/schemas/Pageable"
        size:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
...
    PermissionResponse:
      type: object
      properties:
        authority:
          type: string
        domain:
          type: string
        name:
          type: string
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      name: bearerAuth
      scheme: bearer
      type: http